Split resource chart props
