+++ src/ChartWrapper/ChartWrapper.tsx
@@ -4,18 +4,7 @@ import { Spin } from 'antd';
 import { observer } from 'mobx-react';
 
 import { CategoricalChartWrapper } from 'recharts';
-
-import { IResourceAreaChartProps } from '../AreaChart/ResourceAreaChartProps';
-import { IResourceBarChartProps } from '../BarChart/ResourceBarChartProps';
-import { IResourceComposedChartProps } from '../ComposedChart/ResourceComposedChartProps';
-import { IResourceLineChartProps } from '../LineChart/ResourceLineChartProps';
-import { IResourceRadarChartProps } from '../RadarChart/ResourceRadarChartProps';
-
-export type IResourceChartProps = IResourceAreaChartProps &
-  IResourceBarChartProps &
-  IResourceComposedChartProps &
-  IResourceLineChartProps &
-  IResourceRadarChartProps;
+import { IResourceChartProps } from './ResourceChartProps';
 
 export const chartWrapper = (
   ChartComponent: React.ComponentType<
